fix more odd characters
Note: in UTF-8 0xEF 0xBF 0xBD is a special sequence meaning that the
character isn't supported and is what a conversion program inserts when
it can't work out what to put in. It was evident in spells.csv where
some odd version of a posessive apostraphe had been used.
I loaded the files into Notepad++ and used Find characters in range
non-ASCII to check all the accented characters.