Move the orderchecker data files to orderchecker.jar
This should stop the configuration problems some people were having.
Change orderchecker to load data files from the filesystem in preference
to the jar file, so that users can control customisations.
Change orderchecker to spot when it is packaged in a jar file and load
data files from the jar if the files are not found in the filesystem.
Tweak so that orderchecker will work within joverseer in eclipse.