Fix: Neutral army colour not shown on map.
Also manage the map metadata better in the map renderers. It now uses a
common class AbstractBaseRenderer to cache the mapmetadata. Auto-wired
by spring. refreshConfig() is added to Renderer interface, so that it
can be updated when needed. This decouples the renderers from the mapp
metadata for testing purposes, and speeds up rendering by not having to
check if anything has changed.
Todo: something similar for the preference registry, map options and
color source.