Fix: The order dropdown is back in the OrderListView.
Note that it's behaviour is different between the OrderListView
(normally) bottom-left, and OrderEditor bottom-right.

The OrderEditor *will* only offer orders appropriate to the character.

The OrderListView unfortunately will show all orders - even those that
are wrong for the type of character, e.g. offer an agent-only character
mage orders. The orderchecker process will spot the errors.

This commit also added a simple unit test for the OrderListView, which
forces the minimum java requirement to version 1.8 (for lambda
functions) and adds assertj jars for compiling tests and running tests.