Enh: Merge Export Order Text and Export Order commands.
The two separate export orders were confusing.
The view automatically generates a 'nice' text representation of the
order.
It may be copied to the clipboard.
The Save and Send buttons will use a different format of the orders that
Middle-Earth Games and JOverseer understand.

QA: Developers can now use a GUI designer to tweak the design of the ExportOrderForm.