Order editor and order checker refactoring.
Refactoring so that the various order parmeter edtiors can be used in
the Order List View in the future.

A trial attempt to push the results of orderchecker into a modeless view
so that the information can be reviewed against the other views.

Note: this changes the interface between ordercheck and JO...so a new orderchecker.jar is required.