orderchecker updates
QA: putting the symbolic constants back in for maintainability.
QA: Main is now a dependency of each Rule to allow better testing.
Enh: The Rank Rule now has 2 extra optional parameters :
onlyOne which if non-zero indicates that only one of the same skill is
allowed.
atCapitalReq which if non-zero indicates that the command must be
performed at the capital.