fix: java illegal access
Certain versions of java 9+ causes a warning or a fatal error due to a third-party component (JIDE) accessing private Look And Feel implementation.
As a Permanent Temporary Fix, the --add_export option is required when invoking the java VM.

In adding the --add_export option, it was discovered that some of the batch files had the VM options set *after* the -jar argument, so would have not been applied and broken the passing of .jov files to the .bat files.
The MacOS universal java application stub was also updated from v2.0.2 to v3.2.0