support Bring-Your-Own Java
If you really want to use an existing Java in preference to the supplied one (eg because we didn't include something you need.) then the .bat files have been updated to use an  existing java version when they find a file called 'DontUseCustomJava.txt' in the same place as the .bat file.

If you find that this works for you when the supplied java doesn't work, then please let Middleearth.com know.

The initial known case was an accessibility feature - the onscreen magnifier.