gamesystems changes and versioning policy changes
JOVersion compatibility class added to rename middleearth.com to gamesystems.com
Fix command line parsing which sometimes broke launching JO with a .jov file.
Added -1 option to force version checking routines to run at startup.
Added a new Preference to track the version of JO Compatibility run. At PostStartup if it is not set to the current version, then it runs JOCompatibilty class.
Reduced some of the compiler warnings.
support for versions of the form 1.18.11.alpha and 1.18.11-rc1