1.18.11-rc2
appveyor build tweaked, so that betafeed and latestbetaupdate.zip are built.
No need to rename files when releasing a beta.

A failure to parse the pdf text no longer confusingly hints that it's an XML
problem. Also logs the name of the temporary files used. (breakpoint on cleanup() and ForceReturn in eclipse to leave the temporaries around.

HTTP errors are no longer ignored for meturn submission.